fix VC++ x64 conversion warnings

+++ silicium/win32/directory_changes.cpp
@@ -62,7 +62,7 @@ namespace Si
 				DWORD received = 0;
 				std::array<char, 0x10000> buffer;
 				DWORD const actions = FILE_NOTIFY_CHANGE_FILE_NAME | FILE_NOTIFY_CHANGE_LAST_WRITE | FILE_NOTIFY_CHANGE_DIR_NAME;
-				if (!ReadDirectoryChangesW(watch_file.get(), buffer.data(), buffer.size(), is_recursive, actions, &received, nullptr, nullptr))
+				if (!ReadDirectoryChangesW(watch_file.get(), buffer.data(), static_cast<DWORD>(buffer.size()), is_recursive, actions, &received, nullptr, nullptr))
 				{
 					//TODO: handle ERROR_NOTIFY_ENUM_DIR (overflow of event queue)
 					throw std::logic_error("to do: error handling");
